chore: cleaner types in wallet provider
