+++ packages/contracts/src/submit-tx.ts
@@ -21,7 +21,7 @@ import {
 } from '@midnight-ntwrk/ledger-v6';
 import {
   BALANCE_TRANSACTION_TO_PROVE,
-  type BalancedTransaction,
+  type BalancedProvingRecipe,
   type BalanceTransactionToProve,
   type Contract,
   type FinalizedTxData,
@@ -29,7 +29,6 @@ import {
   NOTHING_TO_PROVE,
   type NothingToProve,
   type ProvenTransaction,
-  type ProvingRecipe,
   TRANSACTION_TO_PROVE,
   type ZKConfig
 } from '@midnight-ntwrk/midnight-js-types';
@@ -85,7 +84,7 @@ function logAndCheckTransaction(circuitId: string | undefined, tx: Transaction<a
   }
 }
 
-async function proveTransaction<C extends Contract, ICK extends ImpureCircuitId<C>>(recipe: ProvingRecipe<BalancedTransaction>, providers: SubmitTxProviders<C, ICK>, proveTxConfig: {
+async function proveTransaction<C extends Contract, ICK extends ImpureCircuitId<C>>(recipe: BalancedProvingRecipe, providers: SubmitTxProviders<C, ICK>, proveTxConfig: {
   zkConfig: ZKConfig<ICK>
 } | undefined) {
   let toSubmit: ProvenTransaction;
