+++ packages/types/src/midnight-types.ts
@@ -23,7 +23,6 @@ import {
   type Transaction,
   type TransactionHash,
   type TransactionId,
-  type UnprovenTransaction
 } from '@midnight-ntwrk/ledger-v6';
 
 /**
@@ -292,28 +291,3 @@ export type UnshieldedBalance = {
  * Represents a collection of unshielded balances, which are balances that are not shielded or encrypted.
  */
 export type UnshieldedBalances = UnshieldedBalance[];
-
-export const TRANSACTION_TO_PROVE = 'TransactionToProve';
-export const BALANCE_TRANSACTION_TO_PROVE = 'BalanceTransactionToProve';
-export const NOTHING_TO_PROVE = 'NothingToProve';
-
-export type TransactionToProve = {
-  readonly type: typeof TRANSACTION_TO_PROVE;
-  readonly transaction: UnprovenTransaction;
-};
-
-export type BalanceTransactionToProve<TTransaction> = {
-  readonly type: typeof BALANCE_TRANSACTION_TO_PROVE;
-  readonly transactionToProve: UnprovenTransaction;
-  readonly transactionToBalance: TTransaction;
-};
-
-export type NothingToProve<TTransaction> = {
-  readonly type: typeof NOTHING_TO_PROVE;
-  readonly transaction: TTransaction;
-};
-
-export type ProvingRecipe<TTransaction> =
-  | TransactionToProve
-  | BalanceTransactionToProve<TTransaction>
-  | NothingToProve<TTransaction>;
