+++ packages/types/src/wallet-provider.ts
@@ -21,9 +21,32 @@ import {
   type UnprovenTransaction,
 } from '@midnight-ntwrk/ledger-v6';
 
-import { type ProvingRecipe } from './midnight-types';
+export const TRANSACTION_TO_PROVE = 'TransactionToProve';
+export const BALANCE_TRANSACTION_TO_PROVE = 'BalanceTransactionToProve';
+export const NOTHING_TO_PROVE = 'NothingToProve';
 
-export type BalancedTransaction = UnprovenTransaction | FinalizedTransaction;
+export type TransactionToProve = {
+  readonly type: typeof TRANSACTION_TO_PROVE;
+  readonly transaction: UnprovenTransaction;
+};
+
+export type BalanceTransactionToProve<TTransaction> = {
+  readonly type: typeof BALANCE_TRANSACTION_TO_PROVE;
+  readonly transactionToProve: UnprovenTransaction;
+  readonly transactionToBalance: TTransaction;
+};
+
+export type NothingToProve<TTransaction> = {
+  readonly type: typeof NOTHING_TO_PROVE;
+  readonly transaction: TTransaction;
+};
+
+export type ProvingRecipe<TTransaction> =
+  | TransactionToProve
+  | BalanceTransactionToProve<TTransaction>
+  | NothingToProve<TTransaction>;
+
+export type BalancedProvingRecipe = ProvingRecipe<UnprovenTransaction | FinalizedTransaction>;
 
 /**
  * Interface representing a WalletProvider that handles operations such as
@@ -37,7 +60,7 @@ export interface WalletProvider {
    * @param newCoins
    * @param ttl
    */
-  balanceTx(tx: UnprovenTransaction, newCoins?: ShieldedCoinInfo[], ttl?: Date): Promise<ProvingRecipe<BalancedTransaction>>;
+  balanceTx(tx: UnprovenTransaction, newCoins?: ShieldedCoinInfo[], ttl?: Date): Promise<BalancedProvingRecipe>;
 
   getCoinPublicKey(): CoinPublicKey;
